+++ packages/core/src/lib/message/topic_only_message.ts
@@ -3,7 +3,7 @@ import type {
   IDecoder,
   IProtoMessage,
 } from "@waku/interfaces";
-import { proto_topic_only_message as proto } from "@waku/proto";
+import { TopicOnlyMessage as ProtoTopicOnlyMessage } from "@waku/proto";
 import debug from "debug";
 
 const log = debug("waku:message:topic-only");
@@ -14,7 +14,7 @@ export class TopicOnlyMessage implements IDecodedMessage {
   public timestamp: undefined;
   public ephemeral: undefined;
 
-  constructor(private proto: proto.TopicOnlyMessage) {}
+  constructor(private proto: ProtoTopicOnlyMessage) {}
 
   get contentTopic(): string {
     return this.proto.contentTopic ?? "";
@@ -25,7 +25,7 @@ export class TopicOnlyDecoder implements IDecoder<TopicOnlyMessage> {
   public contentTopic = "";
 
   fromWireToProtoObj(bytes: Uint8Array): Promise<IProtoMessage | undefined> {
-    const protoMessage = proto.TopicOnlyMessage.decode(bytes);
+    const protoMessage = ProtoTopicOnlyMessage.decode(bytes);
     log("Message decoded", protoMessage);
     return Promise.resolve({
       contentTopic: protoMessage.contentTopic,