+++ packages/message-encryption/src/ecies.ts
@@ -1,10 +1,11 @@
-import { Decoder as DecoderV0, proto } from "@waku/core/lib/message/version_0";
+import { Decoder as DecoderV0 } from "@waku/core/lib/message/version_0";
 import type {
   IDecoder,
   IEncoder,
   IMessage,
   IProtoMessage,
 } from "@waku/interfaces";
+import { WakuMessage } from "@waku/proto";
 import debug from "debug";
 
 import {
@@ -26,7 +27,7 @@ export { DecodedMessage, generatePrivateKey, getPublicKey };
 
 const log = debug("waku:message-encryption:ecies");
 
-class Encoder implements IEncoder {
+export class Encoder implements IEncoder {
   constructor(
     public contentTopic: string,
     private publicKey: Uint8Array,
@@ -38,7 +39,7 @@ class Encoder implements IEncoder {
     const protoMessage = await this.toProtoObj(message);
     if (!protoMessage) return;
 
-    return proto.WakuMessage.encode(protoMessage);
+    return WakuMessage.encode(protoMessage);
   }
 
   async toProtoObj(message: IMessage): Promise<IProtoMessage | undefined> {
@@ -89,7 +90,7 @@ export function createEncoder(
   return new Encoder(contentTopic, publicKey, sigPrivKey, ephemeral);
 }
 
-class Decoder extends DecoderV0 implements IDecoder<DecodedMessage> {
+export class Decoder extends DecoderV0 implements IDecoder<DecodedMessage> {
   constructor(contentTopic: string, private privateKey: Uint8Array) {
     super(contentTopic);
   }