+++ packages/message-encryption/src/symmetric.ts
@@ -1,10 +1,11 @@
-import { Decoder as DecoderV0, proto } from "@waku/core/lib/message/version_0";
+import { Decoder as DecoderV0 } from "@waku/core/lib/message/version_0";
 import type {
   IDecoder,
   IEncoder,
   IMessage,
   IProtoMessage,
 } from "@waku/interfaces";
+import { WakuMessage } from "@waku/proto";
 import debug from "debug";
 
 import {
@@ -25,7 +26,7 @@ export { DecodedMessage, generateSymmetricKey };
 
 const log = debug("waku:message-encryption:symmetric");
 
-class Encoder implements IEncoder {
+export class Encoder implements IEncoder {
   constructor(
     public contentTopic: string,
     private symKey: Uint8Array,
@@ -37,7 +38,7 @@ class Encoder implements IEncoder {
     const protoMessage = await this.toProtoObj(message);
     if (!protoMessage) return;
 
-    return proto.WakuMessage.encode(protoMessage);
+    return WakuMessage.encode(protoMessage);
   }
 
   async toProtoObj(message: IMessage): Promise<IProtoMessage | undefined> {
@@ -87,7 +88,7 @@ export function createEncoder(
   return new Encoder(contentTopic, symKey, sigPrivKey, ephemeral);
 }
 
-class Decoder extends DecoderV0 implements IDecoder<DecodedMessage> {
+export class Decoder extends DecoderV0 implements IDecoder<DecodedMessage> {
   constructor(contentTopic: string, private symKey: Uint8Array) {
     super(contentTopic);
   }