+++ packages/message-encryption/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Export `Encoder` and `Decoder` types.
+
+## [0.0.7] - 2022-12-19
+
+### Fixed
+
+- Incorrect `proto` import.
+
+## [0.0.6] - 2022-12-16
+
+### Fixed
+
+- Type resolution when using `moduleResolution: node`.
+
+## [0.0.5] - 2022-12-15
+
 ### Added
 
 - Add `@multiformats/multiaddr` as peer dependency.
@@ -26,7 +44,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Alpha version of `@waku/message-encryption`.
 
-[unreleased]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.4...HEAD
+[unreleased]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.7...HEAD
+[0.0.7]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.6...@waku/message-encryption@0.0.7
+[0.0.6]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.5...@waku/message-encryption@0.0.6
+[0.0.5]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.4...@waku/message-encryption@0.0.5
 [0.0.4]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.3...@waku/message-encryption@0.0.4
 [0.0.3]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.2...%40waku/message-encryption@0.0.3
 [0.0.2]: https://github.com/waku-org/js-waku/compare/@waku/message-encryption@0.0.1...%40waku/message-encryption@0.0.2